Add support for SRTxP re-keying - part 3

+++ src/main/java/org/tsitle/rtsp/threads/rtsp/RtspConstants.java
@@ -32,7 +32,7 @@ public class RtspConstants {
 	public static final int RTP_SAMPLES_PER_FRAME_AAC_LC_AUDIO_LD = 512;
 
 	/** Interval for re-keying the SRTxP master key and salt (in packets) as integer */
-	public static final int SRTXP_REKEYING_INTERVAL_PACKETS_INT = 1048576;  // ^=2^20
+	public static final long SRTXP_REKEYING_INTERVAL_PACKETS_INT = 1048576;  // ^=2^20
 	/** Interval for re-keying the SRTxP master key and salt (in packets) as exponent of 2 */
 	@SuppressWarnings("unused")
 	public static final String SRTXP_REKEYING_INTERVAL_PACKETS_EXP2_STR = "2^20";  // ^=1048576
